FOG-408: Make ingest peer uri filtering based on responder id (#22)
* FOG-408: Make ingest peer uri filtering based on responder id

This fixes a bug that exists today in fog ingest deployments.
I thought this was only a minor issue because, it would wrongly
add the peer listen uri to the list of peer uris, but then filter
it out and not connect to it. But because the active server replicates
its peer list onto the backups continuously, they will all end up
with this uri in their list, and there is no configuration that will
fix this. Additionally, a server won't be able to filter out it's
own (real) URI that the peers have in their lists, because it doesn't
filter by responder id before this commit.

* fix clippy

* fix unwraps, fix tests

* fix linting

* Make the peers argument parsed as a comma separated list in structopt

* fix structopt warning

* fix struct opt config string

* fix issue with command line arg parsing and add a test