Add cookie support for fog enclave connections
This is meant to mirror this commit from mobilecoin repo:
https://github.com/mobilecoinfoundation/mobilecoin/commit/0e8e7715c4c499b01b85faacbbb4086e3cb90292

onto the fog attested connections.

This makes the generic attested connection object maintain a cookie jar,
and makes the enclave grpc trait take a grpcio::CallOption, and return
grpcio::Metadata objects, as the consensus version does.

This should allow the fog test client to work when load balancing
is in place that requires cookies to be preserved across connections.