fix FOG-180: enclave report cache updated too frequently (#116)
* fix FOG-180: enclave report cache updated too frequently

this commit attempts to make it so that ingest server updates
the enclave report cache only when necessary

This is:
(1) At least once before anything happens
(2) Whenever the ingress key has changed
(3) Whenever we notice that the report doesn't match the key
    we think we are publishing to the database.

Additionally, a background thread is created that updates it
at least as frequently as consensus (18 hours)

Before this commit, we update it at least once per block.

There are some questions about whether DNS failures with looking
for IAS were causing instability in the service.
This commit doesn't speak directly to that, but it
can't hurt to hit up IAS less often than we were doing.

* fixups to previous (mostly around exactly when mutexes are released)

* Move key-extract function to fog-api crate, add test for new_keys call

* add some comments and debug logs

* fix module name

* add missing copyright notice

* simplify fog report parse code to avoid sgx-types dependency (thanks james)

* Add metric for last published report

This is requested in discussions with ops, to make better alerts

Conflicts:
    Cargo.lock