update fog and fog clients for encrypted memo changes: recoverable transaction history (#122)
* WIP update fog and fog clients for encrypted memo changes

This makes fog ingest include the encrypted memo in the TxOutRecord
object. It then passes through fog view as usual. We also update
the FogTxOut object to parse the memo out of the view response.

We don't update the fog sample paykit to actually read the memos,
but it does use the RTH memo build.

This uprevs mobilecoin submodule, and makes the android bindings
and libmobilecoin able to build.

It installs a RTHMemoBuilder when those bindings create transaction
builder, but it does not configure the RTHMemoBuilder, because
we probably need to extend the bindings to do that properly.

It also gets all the tests passing

* fix some crap

* make the test client exercise and test the recoverable transaction history

* add some notes in bindings code

* fix doc comments

* fix fog sample paykit

* fix test client assertion

* add logging to memo handler to help debug CD

* fix memo auth checking and diagnostics

* fix previous

* uprev mobilecoin submodule to point to latest mvp-memo revision

* address review comment

* update for AddressHash -> ShortAddressHash rename

* update for merge of mvp-memo to master